Add serializer demo/test binary
I could run some tests using @gibson042's test suite. It's great!

```
$ git clone git@github.com:gibson042/canonicaljson-spec.git
$ cd canonicaljson-spec/
$ ./test.sh ../canonicaljson-rs/demo/target/debug/demo
```
[output](https://gist.github.com/leplatrem/58d428dae53603843659226348937aff)

The variation of our implementation is larger than I would have
expected, especially around obviously the hardest thing in the world:
numbers! :D

It may perfectly be because I used serde's default JSON reader, since we were
only interested in serialization here. And some `number out of range` :)

I really wish we would be able to align, but it requires study, to guarantee continuity of our service
in production.

We would have to remain within the behaviour our clients implementation in JavaScript.
Yes, because unfortunately at the time we aligned the serialization
of floats on the server after we discovered qwirks, it was indeed simpler and less risky than
breaking clients!

Let's see first if we have had any published data recently
with floats, and how. Maybe forbid them in our system for the sake of
having a single specification and test suite.